update
Також змінила назву папки,З modal_close to modal.js
