+++ src/index.html
@@ -26,7 +26,7 @@
     <include src="./partials-index/modal.html"></include>
     <include src="./partials-index/footer-team.html"></include>
     <!-- <script type="module" src="../node_modules/spin.js/spin.js"></script> -->
-    <script src="../src/js/modal_close.js" type="module"></script>
+    <script src="../src/js/modal.js" type="module"></script>
 
     <script src="./index.js" type="module"></script>
   </body>
