+++ src/library.html
@@ -26,7 +26,7 @@
     <!--Модальні вікна-->
     <include src="./partials-index/modal.html"></include>
 
-    <script src="../src/js/modal_close.js" type="module"></script>
+    <script src="../src/js/modal.js" type="module"></script>
     <!-- <script type="module" src="../node_modules/spin.js/spin.js"></script> -->
     <script src="./library.js" type="module"></script>
   </body>
